+++ nlp/tests/__init__.py
@@ -1,0 +1,5 @@
+import os
+
+CUR_DIR = os.path.dirname(__file__)
+GOLD_FILE = os.path.join(CUR_DIR, "gold_queries.json")
+TEST_FILE = os.path.join(CUR_DIR, "noisy_gold_queries.json")